backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.36
psycopg[binary]>=3.1.0
alembic>=1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
sendgrid==6.11.0

# CORS
fastapi-cors==0.0.6
